refactor: move trpc provider
